Add Win support #3

+++ file_writers.py
@@ -36,9 +36,11 @@ def defaults_config_writer(items, current_type):
     config_file_path = os.path.join(config_folder_path, f"{current_type}.txt")
 
     os.makedirs(config_folder_path, exist_ok=True)
-    with open(config_file_path, "w", encoding="utf-8") as file:
-        for item in items:
-            file.write(item + "\n")
+
+    if not os.path.exists(config_file_path):
+        with open(config_file_path, "w", encoding="utf-8") as file:
+            for item in items:
+                file.write(item + "\n")
 
 
 def file_writer(get_active_window_func):
@@ -61,6 +63,7 @@ def file_writer(get_active_window_func):
             if current_active_app in browsers_names:
                 current_browser_url = {
                     "Google Chrome": get_chrome_url,
+                    "chrome": get_chrome_url,
                     "Safari": get_safari_url,
                     "Microsoft Edge": get_edge_url
                 }.get(current_active_app, lambda: None)()
